+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# upcoming
+
+* Add `ng-templates-paths` helper function
+
 ### 1.0.0
 
 * Add `update-po-files` plugin and basic dev setup